+++ mongoengine/__init__.py
@@ -29,7 +29,7 @@ __all__ = (
 )
 
 
-VERSION = (0, 25, 0)
+VERSION = (0, 26, 0)
 
 
 def get_version():