+++ docs/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 Development
 ===========
 - (Fill this out as you fix issues and develop your features).
+
+Changes in 0.26.0
+=================
 - BREAKING CHANGE: Improved the performance of :meth:`~mongoengine.Document.save()`
   by removing the call to :meth:`~mongoengine.Document.ensure_indexes` unless
   ``meta['auto_create_index_on_save']`` is set to True. With the default settings, Document indexes
@@ -19,7 +22,7 @@ Development
   Although it could work to switch an existing DecimalField to Decimal128Field without applying a migration script,
   it is not recommended to do so (DecimalField uses float/str to store the value, Decimal128Field uses Decimal128).
 - BREAKING CHANGE: When using ListField(EnumField) or DictField(EnumField), the values weren't always cast into the Enum (#2531)
-- BREAKING CHANGE (bugfix) Querying ObjectIdField or ComplexDateTimeField with None will no longer raise a ValidationError (#2681)
+- BREAKING CHANGE (bugfix) Querying ObjectIdField or ComplexDateTimeField with None no longer raise a ValidationError (#2681)
 - Allow updating a field that has an operator name e.g. "type" with .update(set__type="foo"). It was raising an error previously. #2595
 
 Changes in 0.25.0